pyproject.toml
[project]
name = "damin-gambit"
version = "0.1.0"
description = "Local-only agent CLI: natural language to SQLite queries."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "httpx>=0.27",
  "rich>=13.7",
  "SQLAlchemy>=2.0",
  "typer>=0.12",
  "python-dateutil>=2.9",
  "uvicorn[standard]>=0.27",
  "python-jose[cryptography]>=3.3",
]

[project.scripts]
damin-gambit = "damin_gambit.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/damin_gambit"]

